LayoutTests/http/tests/media/modern-media-controls/status-support/status-support-loading-expected.txt

Testing the StatusSupport behavior for the loading state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.statusLabel.text is "Loading"

PASS successfullyParsed is true

TEST COMPLETE
